Placeholders for currying function arguments
To have the ability of postpone passing of some arguments to a currying function I've implemented
placeholder feature. It can be passed instead of any
argument and will result in returning a new function
awaiting those arguments.
I've corrected typos I had in almost every place, it's `currying` not `curring` of course.

Fixes #4